+++ RealmSwift/Object.swift
@@ -80,6 +80,7 @@ import Realm.Private
 
  See our [Swift guide](https://docs.mongodb.com/realm/sdk/swift/fundamentals/relationships/) for more details.
  */
+
 public typealias Object = RealmSwiftObject
 extension Object: _RealmCollectionValueInsideOptional {
     // MARK: Initializers
@@ -103,7 +104,6 @@ extension Object: _RealmCollectionValueInsideOptional {
         self.init()
         RLMInitializeWithValue(self, value, .partialPrivateShared())
     }
-
 
     // MARK: Properties
 
@@ -178,6 +178,38 @@ extension Object: _RealmCollectionValueInsideOptional {
      - returns: An array of property names.
      */
     @objc open class func indexedProperties() -> [String] { return [] }
+
+    /**
+     Override this method to specify a map of public-private property names.
+     This will set a different property column name on the Realm, and allows using the public name
+     for any operation with the property. (Ex: Queries, Sorting, ...).
+     This very helpful if you need to map property names from your `Device Sync` JSON schema
+     to local property names.
+
+     ```swift
+     class Person: Object {
+         @Persisted var firtName: String
+         @Persisted var birthDate: Date
+         @Persisted var age: Int
+
+         override class public func propertiesMapping() -> [String : String] {
+             ["firtName"; "first_name",
+              "lastName"; "last_name"]
+         }
+     }
+     ```
+
+     - note: Only property that have a different column name have to be added to the properties mapping
+     dictionary.
+     - note: This API is enable for both old and modern property declaration syntax only.
+
+     - returns: A dictionary of public-private property names.
+     */
+    @objc open override class func propertiesMapping() -> [String: String] { return [:] }
+
+    /// :nodoc:
+    @available(*, unavailable, renamed: "propertiesMapping", message: "`_realmColumnNames` private API is unavailable in our Swift SDK, please use the override `.propertiesMapping()` instead.")
+    @objc open override class func _realmColumnNames() -> [String: String] { return [:] }
 
     // MARK: Key-Value Coding & Subscripting